Remove MetricsComputatorMixin from ReadDataGrid component
